Cambios en estructura de bd - cambios en código
