+++ modules/rh/views/rh-ausencia/create.php
@@ -14,7 +14,7 @@ $this->params['breadcrumbs'][] = $this->title;
 
     <h1><?= Html::encode($this->title) ?></h1>
 
-    <?= $this->render('_form-wip', [
+    <?= $this->render('_form', [
         'model' => $model,
     ]) ?>
