+++ modules/rh/views/rh-movimiento/index.php
@@ -28,9 +28,10 @@ $this->params['breadcrumbs'][] = $this->title;
             'id',
             'clave_trab',
             'id_plaza',
+            'clave_plaza',
             'id_ausencia',
             'fec_inicio',
-            //'fec_termino',
+            'fec_termino',
             //'tipo_mov',
             //'descr:ntext',
             //'docs:ntext',
