+++ modules/rh/views/rh-movimiento/view.php
@@ -31,13 +31,12 @@ $this->params['breadcrumbs'][] = $this->title;
             'id',
             'clave_trab',
             'id_plaza',
+            'clave_plaza',
             'id_ausencia',
             'fec_inicio',
             'fec_termino',
-            'tipo_mov',
             'descr:ntext',
-            'docs:ntext',
-            'motivo:ntext',
+            'doc:ntext',
             'ref_motivo:ntext',
             'ref_origen:ntext',
         ],
